[buffer-player] Fix play does not resume when change resolution
:Release Notes:
Fix play does not start when change resolution mannually from menu

:Detailed Notes:
After removing load complete timer that was delaying notification
back to chromium pipeline about the init complete, now buffer feed
starts immediately and seek does happen before pipeline load
complete is notified.
So as a result seek call is missed and the player is re-initialized
from chromium side.In this situation pipeline is reciving buffer
whose timestamp is ahead and hence not processed.
Once the seek amount of time is lapsed then playback resumes auto
So if we recieve a ptsToDecode value in load data then we do a seek
for the gst pipeline.

:Testing Performed:
Tested in Raspberry Pi3 board with youtube test app and youtube
site i, enact browser

:QA Notes:

:Issues Addressed:
[PLAT-65386] [webOS OSE]: In YouTube playback based on
    g-media-pipeline, progress bar goes beyond content's end time
[PLAT-65366] Video and audio is stopped though progress bar is
             process  when I change the resolution

Change-Id: I8aff075932910479a935d3ea48e6b5a13c20b291