Fix return type for bufferPlayer test unload function
:Release Notes:
Fix return type for bufferPlayer test unload function

:Detailed Notes:
During g-media-pipeline change for RDX crash issue StreamSource unload
function return type is bool, but as return is handled in if condition
and if control is not taking true path, unload function will not get
return type. This is one of the warning, but some warning can be
converted to Error. Handled the issue introducing return type.

:Testing Performed:
verified locally with bufferPlayer_test

:QA Notes:
N/A

:Issues Addressed:
[WRO-8296] [RDX_CRASH][webos] /usr/sbin/uriplayer_test in
    _ZN3gmp6player17MediaPlayerClientD1Ev (libgmp-player.so + 0x56f74)

Change-Id: I88416a5d9be671a4f54f4cc9c2188e04b20ec11c