Handle Invalid Argument exception for stoll
:Release Notes:
Handle Invalid Argument exception thrown from stoll

:Detailed Notes:
start parameter is parsed and the next string is passed to stoll.
If it is not a valid string with a number content, it throws an invalid
argument error and program crashes. The exception is caught to avoid the crash.

:Testing Performed:
QWTC-475, QWTC-6451, QWTC-402

:QA Notes:
N/A

:Issues Addressed:
[WRO-9224] [RDX_CRASH][webos] /usr/sbin/g-media-pipeline in raise

Change-Id: Ia7ab44127f46967f1025f54b378878b73baaa2f8