[BufferPlayer] Fix live stream playback start issue for webrtc
:Release Notes:
Remove buffer PTS addition for fixing live stream playback issue

:Detailed Notes:
WebRTC remote video buffer PTS is always set to a later time than
the pipeline start. Because of which time stamp for buffers needed
to be adjusted in chromium which leads to more latency.
In this patch set we remove setting time stamp value for buffers
So the video play start immediately and the latency is removed.

:Testing Performed:
Tested live stream playback with change codecs and WebEx conf

:QA Notes:
NA

:Issues Addressed:
[WRP-18882] Fix SW fallback & live stream playback issue

Change-Id: I5575e6dd765760f1e28a128b8449da492e96a284